requirements.txt
python-dotenv=1.*
gspread=5.1*